changed data generation file name and postgres pool details
