+++ database/index.js
@@ -12,7 +12,7 @@
 const { Pool } = require('pg');
 
 const pool = new Pool({
-  user: 'root',
+  user: 'reedbruskin',
   host: 'localhost',
   database: 'fetsy',
 });
